add data ellipse

Adds an ellipse showing the extent of the data to a plot. Most useful for PCA type stuff to show whether the samples separate or not.

Usage

ft_ggplot2_add_ellipse()

Value

a string

See also

Examples

ft_ggplot2_add_ellipse()
#> geom_polygon(stat = "ellipse", aes(fill = aes_fill, color = aes_color))